requirements.txt
aiogram
spotipy
pymongo
python-dotenv
loguru
yt-dlp
mutagen
aiofiles